requirements.txt
# Web framework
Flask>=3.0.0

# Twitter/X API
tweepy>=4.14.0

# Bluesky API
atproto>=0.0.55

# HTTP requests (LinkedIn)
requests>=2.31.0

# LinkedIn OAuth2
requests-oauthlib>=1.3.1

# Unicode grapheme counting (Bluesky uses graphemes, not chars)
grapheme>=0.6.0

# Image processing
Pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
